Add `class CanvasDrawingCurveWithPencilTests`
